Installation script fails on Debian 10
The installation script fails on Debian 10 because the packages list assigned to the variable "software" include "setpriv" which is not available in Debian 10 it was replaced with the package "util-linux", 
so the following line " 856  apt-get -y install $software > /dev/null 2>&1 &" fails quietly without any errors or warnings. it wasn't easy to figure out what the issue.