don't disable show_source (#2746)
i'm guessing it's inclusion is due to some paranoid sysadmin thinking the *name* sounds scary, without even looking at what the function actually does. 
Anyhow, the function is just an alias of highlight_file() , so either both highlight_file() and show_source() should be disabled, or neither of them should be. Also the function is as dangerous as file_get_contents() (: