SFTP support for additional "FTP" accounts (#3844)
* Fix SFTP chroot jail

This handles the addition and deletion of the chroot jail mount points
during all relevant user add, change, and delete operations. It also
modifies sshd configuration to reflect the new chroot jail structure.

* Harden vsftpd by using the new SFTP chroot jail structure

* Revise ssh / sshd service restarts

Debian / Ubuntu now alias sshd service to ssh service so restarting
ssh service is actually restarting it twice so only restart sshd
service. Furthermore rename ssh service restarts to sshd if it is the
only restart command. This works since sshd is the name of the service
on CentOS and Debian / Ubuntu now have aliases for sshd.

* Update upgrade script to enable migration to new chroot jail system

* Remove chroot jail directories when deleting it

* Migrate script

* Fix main.sh format

* Fix some minor issues

* Update ssh config

---------

Co-authored-by: James Buren <braewoods+mgh@braewoods.net>