Update configuration.php (#4827)
* Update configuration.php

Fix disparity between using File Manager as a user with SSH access enabled and with SSH access disabled. When disabled, user was presented with a "home" directory, and when enabled, user was taken straight into their "home/$username" directory. I think this is because of historical code that worked prior to the new sftp jail system, but now seems to result in this disparity. I don't see why any user, with or without SSH or SFTP access, would need to see outside of their own "home/$username" directory via the File Manager?

* Force update Filemanger conf

---------

Co-authored-by: Jaap Marcus <9754650+jaapmarcus@users.noreply.github.com>